MainActivity.java and ConnectionWizard.java are usable
The MainActivity button is not displayed if there is a connection to
the device already. We will need to figure out the name of the
bluetooth module OR we can just code it up to get the value on the
initial connection and then just store it for repeated use.

Read the comment in the ConnectionWizard.java file about the ListView.
It was giving a weird error that I think is fixed.

The last thing that needs to be done for this part is to implement the
client side connection.